backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.13.1

# Validation
jsonschema==4.20.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Payment Processing
stripe==7.11.0

# Email
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==22.0.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
